+++ src/services/account/createAccount.ts
@@ -17,11 +17,3 @@ export async function createAccount(
     },
   });
 }
-
-export async function findAccountsByUserId(userId: string) {
-  return await prisma.account.findMany({
-    where: {
-      userId,
-    },
-  });
-}
