+++ src/services/account/findAccountsByUserId.ts
@@ -1,0 +1,9 @@
+import prisma from "@prisma/index";
+
+export async function findAccountsByUserId(userId: string) {
+  return await prisma.account.findMany({
+    where: {
+      userId,
+    },
+  });
+}
