+++ src/services/auth/signAccessToken.ts
@@ -1,0 +1,14 @@
+import { User } from "@prisma/client";
+import { omit } from "lodash";
+
+import { signJwt } from "@utils/jwt";
+
+export function signAccessToken(user: User) {
+  const payload = omit(user, user.password);
+
+  const accessToken = signJwt(payload, "accessTokenPrivateKey", {
+    expiresIn: "15m",
+  });
+
+  return accessToken;
+}
