+++ src/services/plaid/getPlaidAccounts.ts
@@ -1,0 +1,11 @@
+import { AccountsGetRequest } from "plaid";
+
+import { plaid } from "./plaid";
+
+export async function getPlaidAccounts(accessToken: string) {
+  const request: AccountsGetRequest = {
+    access_token: accessToken,
+  };
+
+  return await plaid.accountsGet(request);
+}
