+++ src/services/user/findUserByEmail.ts
@@ -1,0 +1,9 @@
+import prisma from "@prisma/index";
+
+export async function findUserByEmail(email: string) {
+  return await prisma.user.findUnique({
+    where: {
+      email,
+    },
+  });
+}
