+++ src/services/user/findUserById.ts
@@ -1,0 +1,9 @@
+import prisma from "@prisma/index";
+
+export async function findUserById(id: string) {
+  return await prisma.user.findUnique({
+    where: {
+      id,
+    },
+  });
+}
